modify log_warper 2

+++ include/muradin/base/log_warper.h
@@ -22,7 +22,7 @@ enum	loging_level{
 	ll_end_palce_
 };
 
-typedef void (*OutputFunc)(const char* msg, int len);
+typedef void (*OutputFunc)(const char* msg, size_t len);
 typedef void (*FlushFunc)();
